refactor(layerChooser): migrate locales
